+++ launcher/Application.cpp
@@ -449,6 +449,26 @@ static PyObject *Application_close(PyObject * /*self*/, PyObject *args)
     return PythonSupport::instance()->getNoneReturnValue();
 }
 
+static PyObject *Application_getKeyboardModifiers(PyObject * /*self*/, PyObject *args)
+{
+    Q_UNUSED(args)
+
+    if (qApp->thread() != QThread::currentThread())
+    {
+        PythonSupport::instance()->setErrorString("Must be called on UI thread.");
+        return NULL;
+    }
+
+    bool query = false;
+
+    if (!PythonSupport::instance()->parse()(args, "b", &query))
+        return NULL;
+
+    Qt::KeyboardModifiers modifiers = query ? qApp->queryKeyboardModifiers() : qApp->keyboardModifiers();
+
+    return PythonSupport::instance()->build()("i", int(modifiers));
+}
+
 static PyObject *Application_setQuitOnLastWindowClosed(PyObject * /*self*/, PyObject *args)
 {
     Q_UNUSED(args)
@@ -2102,6 +2122,29 @@ static PyObject *DocumentWindow_setTitle(PyObject * /*self*/, PyObject *args)
         return NULL;
 
     document_window->setWindowTitle(Py_UNICODE_to_QString(title_u));
+
+    return PythonSupport::instance()->getNoneReturnValue();
+}
+
+static PyObject *DocumentWindow_setWindowFilePath(PyObject * /*self*/, PyObject *args)
+{
+    if (qApp->thread() != QThread::currentThread())
+    {
+        PythonSupport::instance()->setErrorString("Must be called on UI thread.");
+        return NULL;
+    }
+
+    PyObject *obj0 = NULL;
+    Py_UNICODE *title_u = NULL;
+    if (!PythonSupport::instance()->parse()(args, "Ou", &obj0, &title_u))
+        return NULL;
+
+    // Grab the document window
+    DocumentWindow *document_window = Unwrap<DocumentWindow>(obj0);
+    if (document_window == NULL)
+        return NULL;
+
+    document_window->setWindowFilePath(Py_UNICODE_to_QString(title_u));
 
     return PythonSupport::instance()->getNoneReturnValue();
 }
@@ -5806,6 +5849,7 @@ static PyMethodDef Methods[] = {
     {"Action_setEnabled", Action_setEnabled, METH_VARARGS, "Action_setEnabled."},
     {"Action_setTitle", Action_setTitle, METH_VARARGS, "Action_setTitle."},
     {"Application_close", Application_close, METH_VARARGS, "Application_close."},
+    {"Application_getKeyboardModifiers", Application_getKeyboardModifiers, METH_VARARGS, "Application_getKeyboardModifiers."},
     {"Application_setQuitOnLastWindowClosed", Application_setQuitOnLastWindowClosed, METH_VARARGS, "Application_setQuitOnLastWindowClosed."},
     {"ButtonGroup_addButton", ButtonGroup_addButton, METH_VARARGS, "ButtonGroup_addButton."},
     {"ButtonGroup_connect", ButtonGroup_connect, METH_VARARGS, "ButtonGroup_connect."},
@@ -5868,6 +5912,7 @@ static PyMethodDef Methods[] = {
     {"DocumentWindow_setPosition", DocumentWindow_setPosition, METH_VARARGS, "DocumentWindow_setPosition."},
     {"DocumentWindow_setSize", DocumentWindow_setSize, METH_VARARGS, "DocumentWindow_setSize."},
     {"DocumentWindow_setTitle", DocumentWindow_setTitle, METH_VARARGS, "DocumentWindow_setTitle."},
+    {"DocumentWindow_setWindowFilePath", DocumentWindow_setWindowFilePath, METH_VARARGS, "DocumentWindow_setWindowFilePath."},
     {"DocumentWindow_setWindowStyle", DocumentWindow_setWindowStyle, METH_VARARGS, "DocumentWindow_setWindowStyle."},
     {"DocumentWindow_show", DocumentWindow_show, METH_VARARGS, "DocumentWindow_show."},
     {"DocumentWindow_tabifyDockWidgets", DocumentWindow_tabifyDockWidgets, METH_VARARGS, "DocumentWindow_tabifyDockWidgets."},